일부 경로 및 사이드바, 푸터 수정
일부 경로 및 사이드바, 푸터 수정
