+++ src/main/java/com/example/service/loginService/LoginService.java
@@ -28,7 +28,6 @@ public interface LoginService {
 	public List<HashMap<String, Object>> getPetListPaging(Pageable paging);
 	public int countPetRecord();
 	
-	
 	/* 로그인 시 비밀번호 유효성검사 */
 	public UserVO checkPass(UserVO vo);
 
@@ -36,12 +35,14 @@ public interface LoginService {
 	public PetVO petAdd(PetVO pvo);
 	void insertPet(String userEmail, PetVO pvo);
 
-	/* 반려견 상세보기 */
-	public PetVO getPetDetail(PetVO pvo);
-	
 	/* 반려견 이미지 등록 */
 	public void insertImgVO(ImgVO ivo);
 	public PetVO getPetOwnerByUser(String userEmail);
+	
+	
+	/* 반려견 상세보기 */
+	public PetVO getPetDetail(PetVO pvo);
+	
 
 	/* 등록한 반려견 불러오기 2 */
 	public List<PetVO> findmMyPet(String userEmail);
