+++ src/main/webapp/WEB-INF/views/friend/friendList.jsp
@@ -107,7 +107,7 @@
 
           <!-- 마이 프로필 -->
           <li class="nav-item">
-            <a href="/include/myPage/myPageProfile" class="nav-link">
+            <a href="/myPage/myPageProfile" class="nav-link">
               <p>
                 마이 프로필
               </p>
@@ -116,7 +116,7 @@
         
           <!-- 나의 게시글 -->
           <li class="nav-item">
-            <a href="/include/myPage/myPageBoard" class="nav-link">
+            <a href="/myPage/myPageBoard" class="nav-link">
               <p>
                 나의 게시글
               </p>
@@ -125,7 +125,7 @@
 
           <!-- 반려동물 -->
           <li class="nav-item">
-            <a href="/include/myPage/myPageDogList" class="nav-link">
+            <a href="/myPage/myPageDogList" class="nav-link">
               <p>
                 반려동물
               </p>
@@ -134,7 +134,7 @@
 
           <!-- 펫친관리 -->
           <li class="nav-item">
-            <a href="/friend/friendList" class="nav-link active">
+            <a href="/friendList" class="nav-link active">
               <p>
                 펫친관리
               </p>
