+++ src/main/webapp/WEB-INF/views/include/Footer.jsp
@@ -10,30 +10,6 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
 
-    <link rel="stylesheet" id="montserrat-css"
-          href="https://fonts.googleapis.com/css?family=Montserrat:300,300i,400,400i,500,500i,600,600i,700,700i,800,800i">
-    <link rel="stylesheet" id="open-sans-css"
-          href="https://fonts.googleapis.com/css?family=Open+Sans:300,300i,400,400i,600,600i,700,700i">
-
-    <link rel="stylesheet" id="animate-css" href="/css/animate.min.css?ver=3.7.2" type="text/css" media="all">
-    <link rel="stylesheet" id="font-awesome-css" href="/css/font-awesome.min.css?ver=5.15.4" type="text/css"
-          media="all">
-    <link rel="stylesheet" id="bootstrap-css" href="/css/bootstrap.min.css?ver=4.6.0" type="text/css" media="all">
-    <link rel="stylesheet" id="smartmenus-bootstrap-css" href="/css/smartmenus-bootstrap.min.css?ver=1.1.1"
-          type="text/css" media="all">
-    <link rel="stylesheet" id="swiper-css" href="/css/swiper.min.css?ver=4.5.3" type="text/css" media="all">
-    <link rel="stylesheet" id="magnific-popup-css" href="/css/magnific-popup.min.css?ver=1.1.0" type="text/css"
-          media="all">
-    <link rel="stylesheet" id="lana-pet-icon-css" href="/css/lana-pet-icon.min.css?ver=1.0.0" type="text/css"
-          media="all">
-    <link rel="stylesheet" id="lana-pet-theme-css" href="/css/lana-pet-theme.min.css?ver=1.0.0" type="text/css"
-          media="all">
-    <link rel="stylesheet" id="lana-pet-print-css" href="/css/lana-pet-print.min.css?ver=1.0.0" type="text/css"
-          media="print">
-    
-    <style type="text/css">
-    
-    </style>
 </head>
 <body class="home page page-template-template-lana-editor">
 
@@ -75,15 +51,7 @@
     </div>
 </footer>
 
-<script type="text/javascript" src="/js/jquery.min.js?ver=3.6.0"></script>
-<script type="text/javascript" src="/js/popper.min.js?ver=1.16.1"></script>
-<script type="text/javascript" src="/js/bootstrap.min.js?ver=4.6.0"></script>
-<script type="text/javascript" src="/js/smartmenus.min.js?ver=1.1.1"></script>
-<script type="text/javascript" src="/js/smartmenus-bootstrap.min.js?ver=1.1.1"></script>
-<script type="text/javascript" src="/js/swiper.min.js?ver=4.5.3"></script>
-<script type="text/javascript" src="/js/scrollmagic.min.js?ver=2.0.8"></script>
-<script type="text/javascript" src="/js/magnific-popup.min.js?ver=1.1.0"></script>
-<script type="text/javascript" src="/js/custom-theme.js?ver=1.0.0"></script>
+
 
 </body>
 </html>
